+++ doc/changelog.md
@@ -13,10 +13,11 @@ Jump to:
 
 Description
 
+- Add EnvironmentConfigLoader for ML Worker Manager
 - Add Model schema with model metadata included
 - Removed device from schemas, MessageHandler and tests
 - Add ML worker manager, sample worker, and feature store
-- Added schemas and MessageHandler class for de/serialization of
+- Add schemas and MessageHandler class for de/serialization of
   inference requests and response messages